Sunshuo/ts bussiness (#4)
* [sunshuo/tsBussiness] modify readme

* support logger

* update modstation

* add source to tss

* add source to tss

* add source to tss

* add session manager

* add session manager

* add

* modify timezone

* format code

* format code

+++ src/ndsec_server/session_manager.h
@@ -15,7 +15,7 @@ public:
    * @brief 获得一个唯一的session句柄
    * @return 32bit大小的数据,即4Byte长的8位十六进制数
    */
-  virtual uint32_t get_session() = 0;
+  virtual uint64_t get_session() = 0;
 
   /**
    * @brief 检查session是否存在/是否正确
@@ -23,7 +23,7 @@ public:
    * @return 是否正确: true  -- 正确或存在与session池中
    *                  false -- 错误或不存在于session池中
    */
-  virtual bool is_session_exist(uint32_t session) = 0;
+  virtual bool is_session_exist(uint64_t session) = 0;
 
   /**
    * @brief 释放单个session句柄
@@ -31,7 +31,7 @@ public:
    * @return 是否成功释放: true  -- 成功释放
    *                     false  -- 释放失败
    */
-  virtual bool free_session(uint32_t session) = 0;
+  virtual bool free_session(uint64_t session) = 0;
 
   /**
    * @brief 清理所有存在的session
